Disable partitions scan for multipathed devices
When multipath devices are being used as disks for VM Guests
any partition scanning / setup should be done within the VM Guest,
not from host. So we need to switch off partitions scanning via
kpartx there.
For this I've implemented a new feature 'no_partitions' which
just serves as a notifier to kpartx to _not_ create partitions
on these devices.

Patch ported to SLES11.

Signed-off-by: Hannes Reinecke <hare@suse.de>